feat: use Material dialogs

+++ grocery-list-client/src/app/grocery-list/grocery-list.component.ts
@@ -4,6 +4,8 @@ import { ApiService } from '../api.service';
 import { GroceryItemComponent } from '../grocery-item/grocery-item.component';
 import { GroceryItem } from 'src/models/grocery-item.type';
 import { MatButtonModule } from '@angular/material/button';
+import { MatDialog, MatDialogModule } from '@angular/material/dialog';
+import { ClearListDialogComponent } from '../clear-list-dialog/clear-list-dialog.component';
 
 @Component({
   selector: 'app-grocery-list',
@@ -12,12 +14,16 @@ import { MatButtonModule } from '@angular/material/button';
     CommonModule,
     GroceryItemComponent,
     MatButtonModule,
+    MatDialogModule,
   ],
   templateUrl: './grocery-list.component.html',
   styleUrls: ['./grocery-list.component.scss']
 })
 export class GroceryListComponent {
-  constructor(private api: ApiService) { }
+  constructor(
+    private api: ApiService,
+    private dialog: MatDialog,
+  ) { }
 
   get groceryList(): GroceryItem[] {
     return this.api.groceryList;
@@ -27,11 +33,12 @@ export class GroceryListComponent {
     return this.groceryList.length === 0;
   }
 
-  // Clear list if user confirms
-  clearList(): void {
-    if (window.confirm(`Warning: This is an irreversible action!
+  openClearListDialog(): void {
+    const dialogRef = this.dialog.open(ClearListDialogComponent);
 
-Please confirm to proceed.`))
-      this.api.deleteGroceryList();
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result)
+        this.api.deleteGroceryList();
+    });
   }
 }
